backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.1.1
pandas==2.2.2
scikit-learn==1.5.2
lightgbm==4.5.0
joblib==1.4.2
PyYAML==6.0.2
pyarrow==17.0.0
google-cloud-bigquery==3.25.0
db-dtypes==1.3.1


